fix(features): remove old files, change name from useMonhltViewNew -> useMonthlyView
